+++ GPS.h
@@ -4,7 +4,7 @@
 # define  PI  3.141592653589793238462643
 
 
-float GPS_getDistance (float currentLong, float currentlat, float destlong, float destlat);
-float GPS_calcTotalDistance(float distance, float *totalDistance);
+double GPS_getDistance (double currentLong, double currentlat, double destlong, double destlat);
+double GPS_calcTotalDistance(double distance, double *totalDistance);
 
 #endif
